+++ app/Http/Controllers/Members/MemberList.php
@@ -3,15 +3,18 @@
 namespace App\Http\Controllers\Members;
 
 use App\Http\Controllers\Controller;
+use App\Models\Configuration;
+use App\Models\ConfigurationKey;
 use App\Models\Filter\MemberFilter;
 use App\Models\Member;
+use App\Models\MemberGroup;
 use Illuminate\Support\Facades\Response;
 
 class MemberList extends Controller
 {
     const CSV_SEPARATOR = ',';
 
-    public function excel()
+    public function excel(): \Symfony\Component\HttpFoundation\StreamedResponse
     {
         return Response::stream(function () {
 
@@ -47,7 +50,7 @@ class MemberList extends Controller
         ]);
     }
 
-    public function csv()
+    public function csv(): \Symfony\Component\HttpFoundation\StreamedResponse
     {
         return Response::stream(function () {
             $file = fopen('php://output', 'w');
@@ -71,37 +74,57 @@ class MemberList extends Controller
         ]);
     }
 
-    private function getBirthdaySortedMembers(bool $allMembers = false): \Illuminate\Database\Eloquent\Collection|array
+    private function getBirthdaySortedMembers(MemberFilter $filer = null): \Illuminate\Database\Eloquent\Collection|array
     {
-        return Member::getAllFiltered(new MemberFilter($allMembers, $allMembers, $allMembers))->whereNotNull('birthday')->get()
+        return Member::getAllFiltered($filer)->whereNotNull('birthday')->get()
             ->sort(function ($memberA, $memberB) {
                 return strcmp($memberA->birthday->isoFormat('MM-DD'), $memberB->birthday->isoFormat('MM-DD'));
             });
     }
 
-    public function birthdayListPrint(bool $allMembers = true)
+    public function birthdayListPrint(): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application
     {
-        $missingBirthdayList = Member::getAllFiltered(new MemberFilter($allMembers, $allMembers, $allMembers))->whereNull('birthday')
+        $filer = $this->getFilterFromConfig();
+        $missingBirthdayList = Member::getAllFiltered($filer)->whereNull('birthday')
             ->orderBy('lastname')->get();
-        $memberList = $this->getBirthdaySortedMembers($allMembers);
+        $memberList = $this->getBirthdaySortedMembers($filer);
 
         return view('members.member-birthday-list-print', [
             'missingBirthdayList' => $missingBirthdayList,
             'members' => $memberList,
-        ]
-        );
+        ]);
     }
 
-    public function birthdayList(bool $allMembers = true)
+    private function getFilterFromConfig(): MemberFilter
     {
-        $missingBirthdayList = Member::getAllFiltered(new MemberFilter($allMembers, $allMembers, $allMembers))->whereNull('birthday')
+        $filterMemberGroup = null;
+        $filterShowPaused = Configuration::getBool(
+            ConfigurationKey::MEMBER_FILTER_SHOW_PAUSED, auth()->user(), false);
+        $filterShowAfterRetired = Configuration::getBool(
+            ConfigurationKey::MEMBER_FILTER_SHOW_AFTER_RETIRED, auth()->user(), false);
+        $filterShowBeforeEntrance = Configuration::getBool(
+            ConfigurationKey::MEMBER_FILTER_SHOW_BEFORE_ENTRANCE, auth()->user(), false);
+
+        $filterMemberGroupId = Configuration::getInt(
+            ConfigurationKey::MEMBER_FILTER_GROUP_ID, auth()->user());
+        if ($filterMemberGroupId) {
+            $filterMemberGroup = MemberGroup::query()->find($filterMemberGroupId)->first();
+        }
+
+        return new MemberFilter($filterShowBeforeEntrance, $filterShowAfterRetired, $filterShowPaused, $filterMemberGroup);
+    }
+
+    public function birthdayList(): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application
+    {
+        $filter = $this->getFilterFromConfig();
+        $missingBirthdayList = Member::getAllFiltered($filter)
+            ->whereNull('birthday')
             ->orderBy('lastname')->get();
-        $memberList = $this->getBirthdaySortedMembers($allMembers);
+        $memberList = $this->getBirthdaySortedMembers($filter);
 
         return view('members.member-birthday-list', [
             'missingBirthdayList' => $missingBirthdayList,
             'members' => $memberList,
-        ]
-        );
+        ]);
     }
 }