+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - use standard notification message for click on calendar link on dashboard
 - store member filter selection in database instead of browser
 - if members are in multiple groups the other components get now update if one changes in attendance record
+- use member filter config also for birthday list (incl print)
 
 ## v0.6.4 - 2024-02-11
 ### Added